Completed Preparing the Data and Creating Chroma DB
Following along in the video, these are the completed sections of "Preparing the Data" and "Creating the Chroma DB". Created the data folder to hold a md version of the Alice's Adventures in Wonderland I got from online and then parsed the data to chunks. Then using those chunks saved it to a new chroma db. Main changes made include changing to local embeddings using sentence-transformers due to OpenAI APIs requirement for credits.